fix(atlassian): resolve FIX-001, FIX-003, FIX-007, FIX-008, FIX-009 from review
- FIX-001: split renderByFormat into render-by-format.ts (decision-only pick-body-format.ts)
- FIX-003: open advanced accordion by default with "API Version" hint in summary
- FIX-007: reference markdown-parsing fractal entry point in INTENT.md Dependencies
- FIX-008: add Confluence v3 silent-fallback intent comment in pick-body-format.ts
- FIX-009: rewrite api_version_override help text in plain Korean+English

Deferred (debt records under .filid/debt/): FIX-002, FIX-004, FIX-005, FIX-006.
