+++ packages/atlassian/src/mcp/pages/setup/index.html
@@ -126,8 +126,8 @@
                 <span class="field-error" hidden></span>
               </div>
 
-              <details class="advanced-accordion">
-                <summary class="advanced-toggle">Advanced</summary>
+              <details class="advanced-accordion" open>
+                <summary class="advanced-toggle">Advanced · API Version</summary>
                 <div class="advanced-body">
                   <div class="form-group checkbox-group">
                     <label>
@@ -171,9 +171,8 @@
                       </label>
                     </div>
                     <p class="form-help">
-                      Server/DC defaults to v2. <br />
-                      Choose v3 only if your DC supports it and you need ADF
-                      body format.
+                      기본값(Auto / v2)은 대부분의 Jira Server·Data Center에서 작동합니다. <br />
+                      v3는 Jira Data Center 8.4 이상에서 ADF(리치 텍스트) 형식이 필요한 경우에만 선택하세요.
                     </p>
                   </div>
                 </div>
