+++ packages/atlassian/src/mcp/tools/fetch/utils/convert-body-for-create.ts
@@ -1,4 +1,5 @@
-import { pickBodyFormat, renderByFormat } from './pick-body-format.js';
+import { pickBodyFormat } from './pick-body-format.js';
+import { renderByFormat } from './render-by-format.js';
 
 /** Convert markdown body for POST (create). Selects ADF/Storage/Wiki by service+apiVersion. */
 export function convertBodyForCreate(
