+++ packages/atlassian/src/mcp/tools/fetch/utils/pick-body-format.ts
@@ -1,5 +1,3 @@
-import { markdownToAdf, markdownToStorage, markdownToWiki } from '../../../../converter/index.js';
-
 export type BodyFormat = 'adf' | 'storage' | 'wiki';
 
 /** Decide the wire body format from the resolved service + API version. */
@@ -7,15 +5,7 @@ export function pickBodyFormat(
   service: 'jira' | 'confluence',
   apiVersion: '2' | '3',
 ): BodyFormat {
+  // Confluence는 현재 storage API만 지원 (v3 미지원). v3 도입 시 분기 추가.
   if (service === 'confluence') return 'storage';
   return apiVersion === '3' ? 'adf' : 'wiki';
-}
-
-/** Render a markdown string into the body format selected by `pickBodyFormat`. */
-export function renderByFormat(md: string, fmt: BodyFormat): unknown {
-  if (fmt === 'adf') return markdownToAdf(md);
-  if (fmt === 'storage') {
-    return { storage: { value: markdownToStorage(md), representation: 'storage' } };
-  }
-  return markdownToWiki(md);
 }
